research_platform/requirements.txt
# Core Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Agent Orchestration
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langchain-google-genai==0.0.5
llama-index==0.9.32
openai==1.7.1
anthropic==0.8.1
google-generativeai==0.3.2

# Paper Sources & APIs
arxiv==2.1.0
pymed==0.8.9
scholarly==1.7.11
semanticscholar==0.4.0
crossref-commons==0.0.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4
pdfminer.six==20221105
pymupdf==1.23.8

# Vector DB & Storage
pinecone-client==3.0.0
weaviate-client==3.26.0
chromadb==0.4.22
qdrant-client==1.7.0
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching & Queue
redis==5.0.1
celery==5.3.4

# ML & Fine-tuning
torch==2.1.2
transformers==4.36.2
peft==0.7.1
accelerate==0.25.0
datasets==2.16.1
scikit-learn==1.3.2
sentence-transformers==2.2.2

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1
tenacity==8.2.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# CLI
click==8.1.7
rich==13.7.0
typer==0.9.0

# Citation Management
bibtexparser==1.4.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
